+++ packages/components/src/components/SearchField/stories/Default.stories.tsx
@@ -5,6 +5,7 @@ import { action } from "storybook/actions";
 import { Label } from "@/components/Label";
 import FieldDescription from "@/components/FieldDescription";
 import { FieldError } from "@/components/FieldError";
+import { Kbd } from "@/components/Kbd";
 
 const meta: Meta<typeof SearchField> = {
   title: "Form Controls/SearchField",
@@ -37,6 +38,14 @@ export const WithFieldError: Story = {
     <SearchField {...props} defaultValue="test" isInvalid>
       <Label>Search</Label>
       <FieldError>Invalid search value</FieldError>
+    </SearchField>
+  ),
+};
+
+export const WithKbd: Story = {
+  render: (props) => (
+    <SearchField {...props}>
+      <Kbd keys={["mod", "k"]} />
     </SearchField>
   ),
 };